requirements.txt
requests
prettytable
python-dotenv
